+++ __tests__/components/needs_selector_test.js
@@ -63,47 +63,4 @@ describe("NeedsSelector", () => {
     props.t = () => "en";
     expect(mountedNeedsSelector());
   });
-
-  it("is expanded if pageWidth > 600px", () => {
-    expect(
-      mountedNeedsSelector()
-        .find("ExpansionPanel")
-        .prop("expanded")
-    ).toEqual(true);
-  });
-
-  it("is not expanded if pageWidth < 600px", () => {
-    props.pageWidth = 100;
-    expect(
-      mountedNeedsSelector()
-        .find("ExpansionPanel")
-        .prop("expanded")
-    ).toEqual(false);
-  });
-
-  it("has no clear button if selectedNeeds is empty", () => {
-    reduxData.selectedNeeds = {};
-    props.store = mockStore(reduxData);
-    expect(
-      mount(<NeedsSelector {...props} {...reduxData} />)
-        .find("#ClearFilters")
-        .first().length
-    ).toEqual(0);
-  });
-
-  it("has a clear button if selectedNeeds is populated", () => {
-    reduxData.selectedNeeds = { foo: "bar" };
-    props.store = mockStore(reduxData);
-    expect(
-      mount(<NeedsSelector {...props} {...reduxData} />)
-        .find("#ClearFilters")
-        .first().length
-    ).toEqual(1);
-  });
-
-  it("has a correct clearNeeds function", () => {
-    let needsInstance = mountedNeedsSelector().instance();
-    needsInstance.clearNeeds();
-    expect(needsInstance.props.setSelectedNeeds).toBeCalledWith({});
-  });
 });
